+++ packages/ag-grid/index.js
@@ -1,0 +1,4 @@
+/**
+ * Lattice Gauge Component
+ */
+module.exports = require('./dist/cjs/index.js');