+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Addition Phase - 2019-12-04
 - Added Map Package.
+- Added ag-grid package.
 
 ## Update MUI version
 - Updated MUI version in dag package to 4.5.1.
@@ -27,6 +28,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Tree
 - Widgets
 - Dag-in-Action Example
+- Lattice-SSR Example
+
 
 ## Update Phase - 2019-10-18
 - Updated minimal example with jest and lint configurations.